+++ settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'todo-ws'
+rootProject.name = 'todo-service'
