+++ src/main/java/net/leanstacks/todosvc/TodoServiceApplication.java
@@ -1,13 +1,13 @@
-package net.leanstacks.todo;
+package net.leanstacks.todosvc;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class TodoApplication {
+public class TodoServiceApplication {
 
 	public static void main(String[] args) {
-		SpringApplication.run(TodoApplication.class, args);
+		SpringApplication.run(TodoServiceApplication.class, args);
 	}
 
 }
